+++ lib/clientUtils.go
@@ -1,7 +1,6 @@
 package lib
 
 import (
-	"context"
 	"encoding/hex"
 	"errors"
 	"io"
@@ -16,14 +15,11 @@ import (
 	"github.com/duke-git/lancet/v2/cryptor"
 	"github.com/duke-git/lancet/v2/fileutil"
 	"github.com/duke-git/lancet/v2/strutil"
-	"github.com/duke-git/lancet/v2/validator"
-	"github.com/hashicorp/go-getter"
 	"github.com/shirou/gopsutil/v3/process"
 	"github.com/spf13/viper"
 	"github.com/unclesp1d3r/cipherswarm-agent-sdk-go/models/components"
 	"github.com/unclesp1d3r/cipherswarm-agent-sdk-go/models/operations"
 	"github.com/unclesp1d3r/cipherswarmagent/lib/arch"
-	"github.com/unclesp1d3r/cipherswarmagent/lib/utils"
 	"github.com/unclesp1d3r/cipherswarmagent/shared"
 )
 
@@ -254,103 +250,6 @@ func writeResponseToFile(responseStream io.Reader, filePath string) error {
 
 	if _, err := io.Copy(file, responseStream); err != nil {
 		return logAndSendError("Error writing hashlist file", err, operations.SeverityCritical, nil)
-	}
-
-	return nil
-}
-
-// downloadFile downloads a file from a given URL and saves it to the specified path with optional checksum verification.
-// If the URL is invalid, it returns an error. If the file already exists and the checksum matches, the download is skipped.
-func downloadFile(url string, path string, checksum string) error {
-	if !validator.IsUrl(url) {
-		shared.Logger.Error("Invalid URL", "url", url)
-
-		return errors.New("invalid URL")
-	}
-
-	if fileExistsAndValid(path, checksum) {
-		shared.Logger.Info("Download already exists", "path", path)
-
-		return nil
-	}
-
-	displayDownloadFile(url, path)
-	if err := downloadAndVerifyFile(url, path, checksum); err != nil {
-		return err
-	}
-	displayDownloadFileComplete(url, path)
-
-	return nil
-}
-
-// fileExistsAndValid checks if a file exists at the given path and, if a checksum is provided, verifies its validity.
-// The function returns true if the file exists and matches the given checksum, or if no checksum is provided.
-// If the file does not exist or the checksum verification fails, appropriate error messages are logged.
-func fileExistsAndValid(path string, checksum string) bool {
-	if !fileutil.IsExist(path) {
-		return false
-	}
-
-	if strutil.IsBlank(checksum) {
-		return true
-	}
-
-	fileChecksum, err := cryptor.Md5File(path)
-	if err != nil {
-		shared.Logger.Error("Error calculating file checksum", "path", path, "error", err)
-
-		return false
-	}
-
-	if fileChecksum == checksum {
-		return true
-	}
-
-	shared.Logger.Warn("Checksums do not match", "path", path, "url_checksum", checksum, "file_checksum", fileChecksum)
-	SendAgentError("Resource "+path+" exists, but checksums do not match", nil, operations.SeverityInfo)
-	if err := os.Remove(path); err != nil {
-		SendAgentError(err.Error(), nil, operations.SeverityMajor)
-		shared.Logger.Error("Error removing file with mismatched checksum", "path", path, "error", err)
-	}
-
-	return false
-}
-
-// downloadAndVerifyFile downloads a file from the given URL and saves it to the specified path, verifying the checksum if provided.
-// If a checksum is given, it is appended to the URL before download. The function then configures a client for secure file transfer.
-// The file is downloaded using the configured client. After downloading, the file's checksum is verified, if provided, to ensure integrity.
-// If the checksum does not match, an error is returned, indicating the downloaded file is corrupt.
-func downloadAndVerifyFile(url string, path string, checksum string) error {
-	if strutil.IsNotBlank(checksum) {
-		var err error
-		url, err = appendChecksumToURL(url, checksum)
-		if err != nil {
-			return err
-		}
-	}
-
-	client := &getter.Client{
-		Ctx:      context.Background(),
-		Dst:      path,
-		Src:      url,
-		Pwd:      shared.State.CrackersPath,
-		Insecure: true,
-		Mode:     getter.ClientModeFile,
-	}
-
-	_ = client.Configure(
-		getter.WithProgress(utils.DefaultProgressBar),
-		getter.WithUmask(os.FileMode(0o022)),
-	)
-
-	if err := client.Get(); err != nil {
-		shared.Logger.Debug("Error downloading file", "error", err)
-
-		return err
-	}
-
-	if strutil.IsNotBlank(checksum) && !fileExistsAndValid(path, checksum) {
-		return errors.New("downloaded file checksum does not match")
 	}
 
 	return nil
